#2 Merging wiki text with old, simpler README
We needed a full documentation on the README file. To reuse what we already had,
we merged the text from the wiki home page
(https://bitbucket.org/brandizzi/golem/wiki/Home) with the older, smaller README
and the so called golem.txt file. Also we renamed README into README.rst because
now we are going to use ReStructuredText.

More work is needed, e.g. turning the README into a doctest file.